feat(0109): GREEN — G1/G5/G6/G8 citation null model drivers
Implements four new permutation drivers for the citation channel null
model. Extracted into _permutation_citation.py to keep compute_null_model.py
under 800 lines. Fixes directed-graph requirement for in-degree metrics
(entropy, PA exponent, betweenness) via _build_union_digraph. G8 caps
n_perm at 50 for tractability.

Also fixes test_zoo_null_ci to match actual z_threshold_upper/lower API.
